requirements.txt
ezdxf
tqdm
aspose-cad